requirements.txt
tensorflow_gpu==1.15.2
numpy==1.18.2
scipy==1.4.1
six
packaging
scikit_image==0.15.0
pandas==0.23.4
matplotlib==3.0.2
Pillow==7.1.2
scikit_learn==0.22


